refactor(upskill): extract registry backends and GitHub operations into
Agent-Id: agent-714b7a26-7577-4ace-86bf-150e9d670d24
Linked-Note-Id: be41d0bd-4809-4294-8fe8-d2e6268a1103
