+++ packages/webapp/src/shell/supplemental-commands/upskill-command.ts
@@ -8,12 +8,10 @@
  * (CORS bypass via host_permissions).
  */
 
-import { unzipSync } from 'fflate';
 import type { Command, CommandContext, SecureFetch } from 'just-bash';
 import { defineCommand } from 'just-bash';
 import type { BrowserAPI, PageInfo } from '../../cdp/index.js';
 import type { VirtualFS } from '../../fs/index.js';
-import { VirtualFS as SharedVirtualFS } from '../../fs/index.js';
 import {
   extractHandoff,
   isSafeUpskillBranch,
@@ -21,8 +19,7 @@ import {
   UPSKILL_REL,
 } from '../../net/handoff-link.js';
 import { parseLinkHeader } from '../../net/link-header.js';
-import { consumeCachedBinaryByUrl } from '../binary-cache.js';
-import { decodeFetchBody, getFetchBodyBytes, parseFetchJson } from '../fetch-body.js';
+import { parseFetchJson } from '../fetch-body.js';
 import {
   buildInstallCmd,
   getInstalledSkillNames,
@@ -32,24 +29,32 @@ import {
   scoreSkills,
 } from './upskill/catalog/catalog.js';
 import { fetchCompanyCatalog, fetchGlobalCatalog } from './upskill/catalog/catalog-fetch.js';
-import { describeFetchError } from './upskill/fetch-error.js';
+import { _resetGlobalFsCache, createGitHubRequestContext } from './upskill/github/github-auth.js';
+import {
+  installFromGitHub,
+  listGitHubSkills,
+  parseGitHubRef,
+} from './upskill/github/github-install.js';
+import { fetchRepoZip, stripZipPrefix } from './upskill/github/github-zip.js';
 import {
   formatDiscoveredSkills,
   formatDiscoveryScope,
   formatSkillInfo,
   upskillHelp,
 } from './upskill/help.js';
+import { installSkillFromZip, runPostInstallHooks } from './upskill/install-pipeline.js';
 import {
-  installSkillFromZip,
-  refreshSprinklesAfterInstall,
-  reloadSkillsAfterInstall,
-  runPostInstallHooks,
-} from './upskill/install-pipeline.js';
+  _resetBrowseShCatalogCache,
+  fetchBrowseShCatalog,
+  installFromBrowseSh,
+  normalizeHostname,
+  parseBrowseShRef,
+} from './upskill/registries/browse-sh.js';
+import { searchRegistries } from './upskill/registries/search.js';
+import { resolveTesslRef } from './upskill/registries/tessl.js';
 import type {
-  BrowseShDetail,
   BrowseShSkillSummary,
   CatalogSkill,
-  GitHubFetchResponse,
   GitHubRequestContext,
   ParsedUpskillFlags,
   RemoteCatalogRow,
@@ -57,1028 +62,21 @@ import type {
   TabCatalogMatch,
   TabUpskillLink,
   TabUpskillResult,
-  TesslSearchResponse,
-  UnifiedSearchResult,
   UserProfile,
 } from './upskill/types.js';
-import {
-  BROWSE_SH_API,
-  GITHUB_API_ACCEPT,
-  GITHUB_GLOBAL_DB,
-  GITHUB_TOKEN_PATH,
-  SKILL_CATALOG_URL,
-  SKILLS_DIR,
-  TESSL_API,
-} from './upskill/types.js';
+import { SKILL_CATALOG_URL } from './upskill/types.js';
 
 export type { BrowseShSkillSummary, TabCatalogMatch, TabUpskillLink, TabUpskillResult };
-// ── Re-exports (preserve the monolith's public surface during the Wave 1 split) ──
-export { scoreSkills };
-
-interface GitHubContent {
-  name: string;
-  path: string;
-  type: 'file' | 'dir';
-  download_url?: string;
-}
-
-interface GitHubErrorBody {
-  message?: string;
-  documentation_url?: string;
-}
-
-let cachedGlobalFsPromise: Promise<VirtualFS> | undefined;
-
-function getGlobalFs(): Promise<VirtualFS> {
-  if (!cachedGlobalFsPromise) {
-    cachedGlobalFsPromise = SharedVirtualFS.create({ dbName: GITHUB_GLOBAL_DB });
-  }
-  return cachedGlobalFsPromise;
-}
-
-/** @internal Exported only for test cleanup. */
-export function _resetGlobalFsCache(): void {
-  const pending = cachedGlobalFsPromise;
-  cachedGlobalFsPromise = undefined;
-  if (!pending) return;
-  // Fire-and-forget dispose so the cached VirtualFS releases its
-  // LightningFS lock (held via navigator.locks). Without this, the
-  // dangling lock request rejects with AbortError on process teardown
-  // and surfaces as an unhandled rejection in tests. Errors during
-  // dispose are intentionally swallowed — this path runs from test
-  // teardown and hot-reload where surfacing a cleanup rejection
-  // produces false failures.
-  pending.then(
-    (vfs) => {
-      void vfs.dispose().catch(() => {});
-    },
-    () => {}
-  );
-}
-
-async function loadConfiguredGitHubToken(): Promise<string | undefined> {
-  try {
-    const globalFs = await getGlobalFs();
-    const token = (await globalFs.readTextFile(GITHUB_TOKEN_PATH)).trim();
-    return token || undefined;
-  } catch {
-    return undefined;
-  }
-}
-
-function buildGitHubHeaders(
-  token?: string,
-  accept: string = GITHUB_API_ACCEPT
-): Record<string, string> {
-  const headers: Record<string, string> = {
-    Accept: accept,
-    'User-Agent': 'slicc-upskill',
-  };
-  if (token) {
-    headers.Authorization = `Bearer ${token}`;
-  }
-  return headers;
-}
-
-async function createGitHubRequestContext(fetch: SecureFetch): Promise<GitHubRequestContext> {
-  const token = await loadConfiguredGitHubToken();
-  return {
-    hasToken: Boolean(token),
-    request: async (url: string, accept: string = GITHUB_API_ACCEPT) => {
-      try {
-        return await fetch(url, {
-          headers: buildGitHubHeaders(token, accept),
-        });
-      } catch (err) {
-        throw new Error(describeFetchError(err, url));
-      }
-    },
-  };
-}
-
-function getHeader(headers: Record<string, string> | undefined, name: string): string | undefined {
-  if (!headers) return undefined;
-  const target = name.toLowerCase();
-  for (const [key, value] of Object.entries(headers)) {
-    if (key.toLowerCase() === target) return value;
-  }
-  return undefined;
-}
-
-function getGitHubErrorDetail(body: Uint8Array | string): string | undefined {
-  const text = decodeFetchBody(body);
-  if (!text) return undefined;
-  try {
-    const parsed = JSON.parse(text) as GitHubErrorBody;
-    if (typeof parsed.message === 'string' && parsed.message.trim()) {
-      return parsed.message.trim();
-    }
-  } catch {
-    // Not JSON — fall back to a trimmed text preview.
-  }
-
-  const trimmed = text.trim();
-  if (!trimmed) return undefined;
-  return trimmed.slice(0, 200);
-}
-
-function formatGitHubFailure(
-  response: GitHubFetchResponse,
-  resourceLabel: string,
-  hasToken: boolean
-): string {
-  const detail = getGitHubErrorDetail(response.body);
-  const detailSuffix = detail ? ` GitHub said: ${detail}` : '';
-  const retryAfter = getHeader(response.headers, 'retry-after');
-  const rateLimitRemaining = getHeader(response.headers, 'x-ratelimit-remaining');
-  const normalizedDetail = detail?.toLowerCase() ?? '';
-  const isRateLimit =
-    response.status === 429 ||
-    rateLimitRemaining === '0' ||
-    normalizedDetail.includes('rate limit');
-
-  if (isRateLimit) {
-    if (hasToken) {
-      return `GitHub rate-limited access to ${resourceLabel} (HTTP ${response.status}). The configured github.token was used, so retry later${retryAfter ? ` after about ${retryAfter} seconds` : ''}.${detailSuffix}`;
-    }
-    return `GitHub rate-limited anonymous access to ${resourceLabel} (HTTP ${response.status}). This often happens on shared VPNs or corporate egress IPs because unauthenticated GitHub API requests are limited per IP. Configure a token with: git config github.token <PAT>, then retry. You can also retry off VPN or later.${detailSuffix}`;
-  }
-
-  if (response.status === 401) {
-    if (hasToken) {
-      return `GitHub rejected the configured github.token while accessing ${resourceLabel} (HTTP 401). Update it with: git config github.token <PAT>, then retry.${detailSuffix}`;
-    }
-    return `GitHub requires authentication to access ${resourceLabel} (HTTP 401). Configure a token with: git config github.token <PAT>, then retry.${detailSuffix}`;
-  }
-
-  if (response.status === 404) {
-    return `GitHub could not find ${resourceLabel} (HTTP 404). Check the repository, path, and permissions.${detailSuffix}`;
-  }
-
-  if (response.status === 403) {
-    if (hasToken) {
-      return `GitHub denied access to ${resourceLabel} (HTTP 403). Check that your github.token can access this repository or retry later if GitHub is throttling requests.${detailSuffix}`;
-    }
-    return `GitHub denied anonymous access to ${resourceLabel} (HTTP 403). If this repo is public on a shared VPN, you may have hit GitHub's shared IP limit; otherwise the repository or path may require authentication. Configure a token with: git config github.token <PAT>, then retry.${detailSuffix}`;
-  }
-
-  const statusDetail = response.statusText ? ` ${response.statusText}` : '';
-  return `GitHub request for ${resourceLabel} failed (HTTP ${response.status}${statusDetail}).${detailSuffix}`;
-}
-
-/**
- * Extract owner/repo from a GitHub URL.
- */
-function parseGitHubUrl(url: string): { owner: string; repo: string } | null {
-  const match = url.match(/github\.com\/([^/?#]+)\/([^/?#]+)/);
-  if (!match) return null;
-  return { owner: match[1], repo: match[2].replace(/\.git$/, '') };
-}
-
-/**
- * Search Tessl registry for skills, returning unified results.
- */
-async function fetchTesslResults(
-  query: string,
-  fetch: SecureFetch
-): Promise<UnifiedSearchResult[]> {
-  const url = `${TESSL_API}/experimental/search?q=${encodeURIComponent(query)}&contentType=skills&page%5Bsize%5D=20`;
-  let response;
-  try {
-    response = await fetch(url, {
-      headers: { Accept: 'application/json' },
-    });
-  } catch (err) {
-    throw new Error(describeFetchError(err, url));
-  }
-  if (response.status !== 200) throw new Error(`Tessl returned HTTP ${response.status}`);
-  const data = parseFetchJson<TesslSearchResponse>(response.body);
-  if (!data.data) return [];
-
-  // Filter to skills only (exclude tiles), deduplicate by sourceUrl
-  const seen = new Map<string, UnifiedSearchResult>();
-  for (const item of data.data) {
-    if (item.type !== 'skill') continue;
-    const a = item.attributes;
-    const gh = parseGitHubUrl(a.sourceUrl);
-    const repo = gh ? `${gh.owner}/${gh.repo}` : undefined;
-    const score = a.scores.aggregate != null ? Math.round(a.scores.aggregate * 100) : null;
-    const key = a.sourceUrl || item.id;
-    const existing = seen.get(key);
-    // Keep the highest-scored entry per source repo
-    if (
-      existing &&
-      existing.qualityScore != null &&
-      score != null &&
-      existing.qualityScore >= score
-    )
-      continue;
-    // Derive skill directory from path (parent of SKILL.md)
-    const skillDir = a.path.replace(/\/SKILL\.md$/i, '');
-    const skillId = skillDir.split('/').pop() || a.name;
-    const installHint = gh
-      ? `upskill ${gh.owner}/${gh.repo} --path ${skillDir.split('/').slice(0, -1).join('/') || '.'} --skill ${skillId}`
-      : `upskill tessl:${a.name}`;
-    seen.set(key, {
-      name: a.name,
-      displayName: a.name,
-      summary: a.description || '',
-      source: 'tessl' as const,
-      qualityScore: score,
-      installHint,
-      featured: a.featured,
-      sourceRepo: repo,
-    });
-  }
-  return Array.from(seen.values());
-}
-
-// ── browse.sh registry ──
-
-let cachedBrowseShCatalog: BrowseShSkillSummary[] | undefined;
-let cachedBrowseShCatalogPromise: Promise<BrowseShSkillSummary[]> | undefined;
-
-/** @internal Exported only for test cleanup. */
-export function _resetBrowseShCatalogCache(): void {
-  cachedBrowseShCatalog = undefined;
-  cachedBrowseShCatalogPromise = undefined;
-}
-
-/**
- * Fetch the full browse.sh catalog. The list is ~200KB and CORS-open, so a
- * single fetch per shell session is fine — the result is cached in-module
- * for the lifetime of the process. Failures clear the cache so the next call
- * retries.
- */
-export async function fetchBrowseShCatalog(fetch: SecureFetch): Promise<BrowseShSkillSummary[]> {
-  if (cachedBrowseShCatalog) return cachedBrowseShCatalog;
-  if (cachedBrowseShCatalogPromise) return cachedBrowseShCatalogPromise;
-  cachedBrowseShCatalogPromise = (async () => {
-    let response;
-    try {
-      response = await fetch(BROWSE_SH_API, { headers: { Accept: 'application/json' } });
-    } catch (err) {
-      throw new Error(describeFetchError(err, BROWSE_SH_API));
-    }
-    if (response.status !== 200) {
-      throw new Error(`browse.sh returned HTTP ${response.status}`);
-    }
-    const data = parseFetchJson<{ skills?: BrowseShSkillSummary[] } | BrowseShSkillSummary[]>(
-      response.body
-    );
-    const skills = Array.isArray(data) ? data : (data.skills ?? []);
-    cachedBrowseShCatalog = skills;
-    return skills;
-  })();
-  try {
-    return await cachedBrowseShCatalogPromise;
-  } catch (err) {
-    cachedBrowseShCatalogPromise = undefined;
-    throw err;
-  }
-}
-
-/**
- * Search the cached browse.sh catalog and return unified results. Filters
- * client-side against `title`, `name`, `description`, `hostname`, and `tags`.
- */
-async function fetchBrowseShResults(
-  query: string,
-  fetch: SecureFetch
-): Promise<UnifiedSearchResult[]> {
-  const catalog = await fetchBrowseShCatalog(fetch);
-  const q = query.trim().toLowerCase();
-  if (!q) return [];
-
-  const matches = catalog.filter((s) => {
-    const haystack = [
-      s.title ?? '',
-      s.name ?? '',
-      s.description ?? '',
-      s.hostname ?? '',
-      ...(s.tags ?? []),
-    ]
-      .join(' ')
-      .toLowerCase();
-    return haystack.includes(q);
-  });
-
-  return matches.map((s) => ({
-    name: s.slug,
-    displayName: s.title || s.name || s.task || s.slug,
-    summary: s.description || '',
-    source: 'browseSh' as const,
-    qualityScore: null,
-    installHint: `upskill browse:${s.hostname}/${s.task}`,
-    sourceRepo: s.hostname,
-  }));
-}
-
-const BROWSE_SH_SEGMENT_RE = /^[A-Za-z0-9._-]+$/;
-
-/**
- * Reject path-traversal-shaped segments (`.`, `..`) and empty strings. The
- * `BROWSE_SH_SEGMENT_RE` allowlist on its own accepts `.` and `..` as full
- * segments because `.` and `-` are in the character class; this helper closes
- * that gap so neither shorthand refs nor frontmatter-derived names can produce
- * `/workspace/skills/browse-./..-..` style targets.
- */
-function isSafeBrowseShSegment(seg: string): boolean {
-  if (!seg) return false;
-  if (seg === '.' || seg === '..') return false;
-  return BROWSE_SH_SEGMENT_RE.test(seg);
-}
-
-/**
- * Parse a browse.sh reference.
- *
- * Accepts:
- * - `browse:{hostname}/{task}` (shorthand)
- * - `https://browse.sh/skills/{hostname}/{task}` (URL form, trailing slash ok)
- *
- * Each segment must satisfy `[A-Za-z0-9._-]+` AND must not be `.` or `..`
- * (no path traversal, no shell metachars). Hostname is normalized to lowercase
- * with a leading `www.` stripped so refs match the install/match logic
- * elsewhere in this file. Returns null for anything else.
- */
-export function parseBrowseShRef(ref: string): { hostname: string; task: string } | null {
-  let hostnameTask: string | undefined;
-
-  if (ref.startsWith('browse:')) {
-    hostnameTask = ref.slice('browse:'.length);
-  } else {
-    const url = ref.match(/^https:\/\/browse\.sh\/skills\/([^/?#]+)\/([^/?#]+?)\/?$/);
-    if (url) hostnameTask = `${url[1]}/${url[2]}`;
-  }
-  if (!hostnameTask) return null;
-
-  const slash = hostnameTask.indexOf('/');
-  if (slash < 0) return null;
-  const rawHostname = hostnameTask.slice(0, slash);
-  const task = hostnameTask.slice(slash + 1);
-  if (!rawHostname || !task) return null;
-  if (task.includes('/')) return null;
-  if (!isSafeBrowseShSegment(rawHostname) || !isSafeBrowseShSegment(task)) return null;
-  const hostname = normalizeHostname(rawHostname);
-  // normalizeHostname only lowercases + strips a leading `www.`; re-verify the
-  // result still satisfies the segment allowlist so a hostname like `www..`
-  // (which normalizes to `.`) is still rejected.
-  if (!isSafeBrowseShSegment(hostname)) return null;
-  return { hostname, task };
-}
-
-/**
- * Extract a top-level scalar field from minimal YAML frontmatter. Only handles
- * plain `key: value` lines (quoted or unquoted) — arrays / block scalars are
- * not parsed. Returns undefined when missing.
- */
-function extractFrontmatterField(skillMd: string, field: string): string | undefined {
-  const fmMatch = skillMd.match(/^---\r?\n([\s\S]*?)\r?\n---/);
-  if (!fmMatch) return undefined;
-  for (const line of fmMatch[1].split(/\r?\n/)) {
-    const m = line.match(/^([A-Za-z_][A-Za-z0-9_-]*)\s*:\s*(.*)$/);
-    if (!m || m[1] !== field) continue;
-    let value = m[2].trim();
-    if (
-      (value.startsWith('"') && value.endsWith('"')) ||
-      (value.startsWith("'") && value.endsWith("'"))
-    ) {
-      value = value.slice(1, -1);
-    }
-    return value || undefined;
-  }
-  return undefined;
-}
-
-/**
- * Build the SLICC adapter preamble inserted below the upstream frontmatter.
- * Same wording for every browse.sh skill regardless of `recommendedMethod` —
- * only the slug and `updated` date vary per skill.
- */
-function buildBrowseShPreamble(detail: BrowseShDetail, slug: string): string {
-  const updated = detail.updated ? ` · updated ${detail.updated}` : '';
-  return [
-    `> [!NOTE] **Imported from browse.sh** — original slug: \`${slug}\``,
-    `>`,
-    `> **SLICC adaptation:** use \`playwright-cli\` — you are running inside the user's real browser session, so the bot-detection workarounds the upstream skill assumes are usually unnecessary.`,
-    `>`,
-    `> Source: <https://browse.sh/skills/${slug}>${updated}`,
-  ].join('\n');
-}
-
-/**
- * Insert the SLICC adapter preamble immediately below the upstream YAML
- * frontmatter. The upstream frontmatter and body MUST round-trip byte-identical
- * around the preamble — we splice `\n<preamble>\n\n` between the closing `---`
- * fence and whatever bytes followed it.
- */
-function insertBrowseShPreamble(skillMd: string, preamble: string): string {
-  const fmMatch = skillMd.match(/^(---\r?\n[\s\S]*?\r?\n---)(\r?\n|$)/);
-  if (!fmMatch) {
-    // No frontmatter — emit the preamble as the file header so downstream
-    // skill loading still sees the SLICC adaptation note.
-    return `${preamble}\n\n${skillMd}`;
-  }
-  const frontmatter = fmMatch[1];
-  const afterFence = fmMatch[2] || '\n';
-  const rest = skillMd.slice(fmMatch[0].length);
-  return `${frontmatter}${afterFence}\n${preamble}\n\n${rest}`;
-}
-
-/**
- * Install a single browse.sh skill into `/workspace/skills/browse-{hostname}-{name}/`.
- *
- * - GETs the detail endpoint for `skillMd`/`skillMdUrl`.
- * - Prefers the Vercel Blob URL (CORS-safe) for the raw markdown body; falls
- *   back to the inline `skillMd` field if the blob fetch fails or is absent.
- * - Honors `force` for collision overwrites.
- * - Writes a single `SKILL.md` with the SLICC adapter preamble inserted below
- *   the upstream YAML frontmatter.
- */
-async function installFromBrowseSh(
-  hostname: string,
-  task: string,
-  fs: VirtualFS,
-  fetch: SecureFetch,
-  force: boolean = false
-): Promise<{ stdout: string; stderr: string; exitCode: number }> {
-  const slug = `${hostname}/${task}`;
-  const detailUrl = `${BROWSE_SH_API}/${hostname}/${task}`;
-
-  let detail: BrowseShDetail;
-  try {
-    const response = await fetch(detailUrl, { headers: { Accept: 'application/json' } });
-    if (response.status === 404) {
-      return {
-        stdout: '',
-        stderr: `upskill: browse.sh skill "${slug}" not found\n`,
-        exitCode: 1,
-      };
-    }
-    if (response.status !== 200) {
-      return {
-        stdout: '',
-        stderr: `upskill: browse.sh returned HTTP ${response.status} for "${slug}"\n`,
-        exitCode: 1,
-      };
-    }
-    detail = parseFetchJson<BrowseShDetail>(response.body);
-  } catch (err) {
-    const msg = describeFetchError(err, detailUrl);
-    return {
-      stdout: '',
-      stderr: `upskill: failed to fetch browse.sh skill "${slug}": ${msg}\n`,
-      exitCode: 1,
-    };
-  }
-
-  let skillMd: string | undefined;
-  if (detail.skillMdUrl) {
-    try {
-      const blobResponse = await fetch(detail.skillMdUrl, { headers: { Accept: 'text/plain' } });
-      if (blobResponse.status === 200) {
-        skillMd = decodeFetchBody(blobResponse.body);
-      }
-    } catch {
-      // fall through to inline
-    }
-  }
-  if (!skillMd && detail.skillMd) {
-    skillMd = detail.skillMd;
-  }
-  if (!skillMd) {
-    return {
-      stdout: '',
-      stderr: `upskill: browse.sh skill "${slug}" has no SKILL.md content\n`,
-      exitCode: 1,
-    };
-  }
-
-  // Derive install dir name. Prefer `name` parsed from the upstream
-  // frontmatter; fall back to `task` with a trailing `-xxxxxx` suffix
-  // stripped (browse.sh appends a short hash to disambiguate variants).
-  const frontmatterName = extractFrontmatterField(skillMd, 'name');
-  const fallbackName = task.replace(/-[A-Za-z0-9]{4,8}$/, '');
-  const skillName = frontmatterName || fallbackName || task;
-  // `hostname` and `task` here came through `parseBrowseShRef`, but `skillName`
-  // can be sourced from untrusted upstream frontmatter — constrain it to the
-  // same safe-segment allowlist before composing the install path. Reject
-  // anything that could escape `/workspace/skills/` (path separators, `.` /
-  // `..` segments, NUL, shell metachars).
-  if (!isSafeBrowseShSegment(skillName) || skillName.length > 64) {
-    return {
-      stdout: '',
-      stderr: `upskill: refusing to install browse.sh skill with unsafe name "${skillName}"\n`,
-      exitCode: 1,
-    };
-  }
-  // Defense in depth: re-validate the hostname segment too. parseBrowseShRef
-  // already guarantees this, but install paths are sensitive enough that we
-  // shouldn't trust the call site.
-  if (!isSafeBrowseShSegment(hostname)) {
-    return {
-      stdout: '',
-      stderr: `upskill: refusing to install browse.sh skill with unsafe hostname "${hostname}"\n`,
-      exitCode: 1,
-    };
-  }
-  const dirName = `browse-${hostname}-${skillName}`;
-  const destDir = `${SKILLS_DIR}/${dirName}`;
-
-  try {
-    await fs.stat(destDir);
-    if (!force) {
-      return {
-        stdout: '',
-        stderr: `upskill: skill "${dirName}" already exists (use --force to overwrite)\n`,
-        exitCode: 1,
-      };
-    }
-    await fs.rm(destDir, { recursive: true });
-  } catch {
-    // doesn't exist, continue
-  }
-
-  const preamble = buildBrowseShPreamble(detail, slug);
-  const fileContent = insertBrowseShPreamble(skillMd, preamble);
-
-  await fs.mkdir(destDir, { recursive: true });
-  await fs.writeFile(`${destDir}/SKILL.md`, fileContent);
-
-  await runPostInstallHooks();
-
-  return {
-    stdout: `Installed skill "${dirName}" from browse.sh (${slug})\n`,
-    stderr: '',
-    exitCode: 0,
-  };
-}
-
-/**
- * Search registries for skills, merge and paginate results.
- *
- * Structured as a list of registry fetchers so additional backends (e.g.
- * browse.sh) can plug in alongside Tessl without restructuring the merge,
- * pagination, or error-handling logic below.
- */
-const SEARCH_PAGE_SIZE = 10;
-
-interface RegistrySource {
-  label: string;
-  host: string;
-  fetch: (query: string, fetch: SecureFetch) => Promise<UnifiedSearchResult[]>;
-}
-
-const REGISTRY_SOURCES: RegistrySource[] = [
-  { label: 'Tessl', host: new URL(TESSL_API).host, fetch: fetchTesslResults },
-  { label: 'browse.sh', host: new URL(BROWSE_SH_API).host, fetch: fetchBrowseShResults },
-];
-
-/**
- * Round-robin interleave per-source result lists, preserving within-source
- * order. Take the first hit from each source in order, then the second from
- * each, etc., skipping any source that has been exhausted. This gives each
- * registry visibility in the top page of results rather than burying browse.sh
- * behind Tessl (or vice versa).
- */
-function interleaveResults(perSource: UnifiedSearchResult[][]): UnifiedSearchResult[] {
-  const merged: UnifiedSearchResult[] = [];
-  const maxLen = perSource.reduce((m, list) => Math.max(m, list.length), 0);
-  for (let i = 0; i < maxLen; i++) {
-    for (const list of perSource) {
-      if (i < list.length) merged.push(list[i]);
-    }
-  }
-  return merged;
-}
-
-async function searchRegistries(
-  query: string,
-  fetch: SecureFetch,
-  page: number = 1
-): Promise<{ stdout: string; stderr: string; exitCode: number }> {
-  const settled = await Promise.allSettled(REGISTRY_SOURCES.map((src) => src.fetch(query, fetch)));
-
-  const perSource = settled.map((s) => (s.status === 'fulfilled' ? s.value : []));
-  const allFailed = settled.every((s) => s.status === 'rejected');
-  const merged: UnifiedSearchResult[] = interleaveResults(perSource);
-
-  // Per-source visibility: surface each failing registry on stderr so the user
-  // can tell which host went down (e.g. api.tessl.io vs browse.sh) rather than
-  // seeing a single generic "no results" line.
-  let warnings = '';
-  for (let i = 0; i < settled.length; i++) {
-    const s = settled[i];
-    if (s.status === 'rejected') {
-      const src = REGISTRY_SOURCES[i];
-      const msg = s.reason instanceof Error ? s.reason.message : String(s.reason);
-      warnings += `warning: ${src.label} registry unavailable (${src.host}): ${msg}\n`;
-    }
-  }
-
-  if (merged.length === 0) {
-    const stderr = allFailed ? `${warnings}upskill: registries failed to respond\n` : warnings;
-    return {
-      stdout: `No skills found for "${query}"\n\nTry a different search term or browse the registries at https://tessl.io/registry or https://browse.sh\n`,
-      stderr,
-      exitCode: allFailed ? 1 : 0,
-    };
-  }
-
-  const totalResults = merged.length;
-  const totalPages = Math.ceil(totalResults / SEARCH_PAGE_SIZE);
-  const safePage = Math.max(1, Math.min(page, totalPages));
-  const startIdx = (safePage - 1) * SEARCH_PAGE_SIZE;
-  const pageResults = merged.slice(startIdx, startIdx + SEARCH_PAGE_SIZE);
-
-  let output = `Search results for "${query}" (page ${safePage}/${totalPages}, ${totalResults} total):\n\n`;
-
-  for (const skill of pageResults) {
-    const scoreStr = skill.qualityScore != null ? String(skill.qualityScore).padStart(3) : '   ';
-    const tag = `[${skill.source}]`;
-    const repoStr = skill.sourceRepo ? `  ${skill.sourceRepo}` : '';
-    output += `  ${skill.name.padEnd(30)} ${scoreStr} ${tag.padEnd(10)}${repoStr}\n`;
-    if (skill.summary) {
-      output += `    ${skill.summary}\n`;
-    }
-    output += '\n';
-  }
-
-  if (safePage < totalPages) {
-    output += `Showing ${startIdx + 1}-${startIdx + pageResults.length} of ${totalResults}. `;
-    output += `Next page: upskill search ${query} --page ${safePage + 1}\n\n`;
-  }
-
-  output += `To install:\n`;
-  output += `  From Tessl:    upskill <owner/repo> --skill <name>\n`;
-  output += `  From browse.sh: upskill browse:<hostname>/<task>\n`;
-
-  return { stdout: output, stderr: '', exitCode: 0 };
-}
-
-/**
- * Parse Tessl reference (tessl:name) and resolve to GitHub source.
- */
-async function resolveTesslRef(
-  name: string,
-  fetch: SecureFetch
-): Promise<
-  { owner: string; repo: string; skillPath: string; skillName: string } | { error: string }
-> {
-  const url = `${TESSL_API}/experimental/search?q=${encodeURIComponent(name)}&contentType=skills&page%5Bsize%5D=5`;
-  let response;
-  try {
-    response = await fetch(url, {
-      headers: { Accept: 'application/json' },
-    });
-  } catch (err) {
-    return { error: `Tessl search failed: ${describeFetchError(err, url)}` };
-  }
-  if (response.status !== 200) {
-    return { error: `Tessl search failed (HTTP ${response.status})` };
-  }
-  const data = parseFetchJson<TesslSearchResponse>(response.body);
-  // Find exact name match among skills
-  const match = data.data?.find((item) => item.type === 'skill' && item.attributes.name === name);
-  if (!match) {
-    return { error: `skill "${name}" not found on Tessl registry` };
-  }
-  const gh = parseGitHubUrl(match.attributes.sourceUrl);
-  if (!gh) {
-    return { error: `skill "${name}" has no GitHub source URL` };
-  }
-  // Derive skill directory path (parent of SKILL.md)
-  const skillDir = match.attributes.path.replace(/\/SKILL\.md$/i, '');
-  return { owner: gh.owner, repo: gh.repo, skillPath: skillDir, skillName: name };
-}
-
-type ZipResult =
-  | { status: 'ok'; files: Record<string, Uint8Array> }
-  | { status: 'error'; message: string };
-
-/**
- * Download and cache a repo ZIP archive from codeload.github.com (not rate-limited).
- */
-async function fetchRepoZip(
-  owner: string,
-  repo: string,
-  fetch: SecureFetch,
-  branch: string = 'main'
-): Promise<ZipResult> {
-  const url = `https://codeload.github.com/${owner}/${repo}/zip/refs/heads/${branch}`;
-  let response;
-  try {
-    response = await fetch(url, {
-      headers: { 'User-Agent': 'slicc-upskill' },
-    });
-  } catch (err) {
-    return { status: 'error', message: describeFetchError(err, url) };
-  }
-  if (response.status === 404) {
-    // Try 'master' branch as fallback
-    if (branch === 'main') {
-      return fetchRepoZip(owner, repo, fetch, 'master');
-    }
-    return { status: 'error', message: 'codeload returned HTTP 404' };
-  }
-  if (response.status !== 200) {
-    return { status: 'error', message: `codeload returned HTTP ${response.status}` };
-  }
-
-  let zipBytes = consumeCachedBinaryByUrl(url);
-  if (!zipBytes) {
-    zipBytes = getFetchBodyBytes(response.body);
-  }
-
-  try {
-    return { status: 'ok', files: unzipSync(zipBytes) };
-  } catch (e) {
-    return {
-      status: 'error',
-      message: `failed to unzip: ${e instanceof Error ? e.message : String(e)}`,
-    };
-  }
-}
-
-/**
- * Strip the top-level directory prefix from zip entries (e.g. "repo-main/foo" → "foo").
- */
-function stripZipPrefix(files: Record<string, Uint8Array>): Record<string, Uint8Array> {
-  const result: Record<string, Uint8Array> = {};
-  for (const [path, content] of Object.entries(files)) {
-    const slashIdx = path.indexOf('/');
-    if (slashIdx < 0) continue; // top-level entry (the directory itself)
-    const stripped = path.slice(slashIdx + 1);
-    if (stripped) result[stripped] = content;
-  }
-  return result;
-}
-
-/**
- * List skills in a GitHub repository.
- * Tries the codeload ZIP first (not rate-limited), falls back to the Contents API.
- */
-async function listGitHubSkills(
-  owner: string,
-  repo: string,
-  github: GitHubRequestContext,
-  subPath?: string,
-  fetch?: SecureFetch,
-  branch?: string
-): Promise<{ skills: Array<{ name: string; path: string }>; error?: string }> {
-  // Try ZIP-based discovery first (no rate limit)
-  if (fetch) {
-    const zip = await fetchRepoZip(owner, repo, fetch, branch);
-    if (zip.status === 'ok') {
-      const files = stripZipPrefix(zip.files);
-      const skills: Array<{ name: string; path: string }> = [];
-      const prefix = subPath ? subPath.replace(/^\/|\/$/g, '') + '/' : '';
-
-      for (const path of Object.keys(files)) {
-        if (!path.startsWith(prefix)) continue;
-        const basename = path.split('/').pop() || '';
-        if (basename === 'SKILL.md') {
-          const skillPath = path.replace(/\/SKILL\.md$/, '');
-          const skillName = skillPath.split('/').pop() || skillPath;
-          skills.push({ name: skillName, path: skillPath });
-        }
-      }
-      return { skills };
-    }
-    // zip.status === 'error' — fall through to API
-  }
-
-  // Fallback: Contents API (rate-limited for anonymous users)
-  const skills: Array<{ name: string; path: string }> = [];
-
-  async function scanDir(path: string): Promise<void> {
-    const base = `https://api.github.com/repos/${owner}/${repo}/contents/${path}`;
-    const url = branch ? `${base}?ref=${encodeURIComponent(branch)}` : base;
-    const response = await github.request(url);
-
-    if (response.status !== 200) {
-      throw new Error(
-        formatGitHubFailure(response, `${owner}/${repo}${path ? `/${path}` : ''}`, github.hasToken)
-      );
-    }
-
-    const contents = parseFetchJson<GitHubContent[]>(response.body);
-
-    for (const item of contents) {
-      if (item.type === 'file' && item.name === 'SKILL.md') {
-        const skillPath = item.path.replace('/SKILL.md', '');
-        const skillName = skillPath.split('/').pop() || skillPath;
-        skills.push({ name: skillName, path: skillPath });
-      } else if (item.type === 'dir') {
-        await scanDir(item.path);
-      }
-    }
-  }
-
-  try {
-    await scanDir(subPath || '');
-    return { skills };
-  } catch (err) {
-    const msg = err instanceof Error ? err.message : String(err);
-    return { skills: [], error: msg };
-  }
-}
-
-async function downloadGitHubDir(
-  items: GitHubContent[],
-  destBase: string,
-  owner: string,
-  repo: string,
-  branch: string | undefined,
-  fs: VirtualFS,
-  github: GitHubRequestContext
-): Promise<void> {
-  for (const item of items) {
-    if (item.type === 'file' && item.download_url) {
-      const fileResponse = await github.request(item.download_url, '*/*');
-      if (fileResponse.status !== 200) {
-        throw new Error(
-          formatGitHubFailure(fileResponse, `${owner}/${repo}/${item.path}`, github.hasToken)
-        );
-      }
-      const cached = consumeCachedBinaryByUrl(item.download_url);
-      await fs.writeFile(`${destBase}/${item.name}`, cached ?? fileResponse.body);
-    } else if (item.type === 'dir') {
-      const subBase = `https://api.github.com/repos/${owner}/${repo}/contents/${item.path}`;
-      const subUrl = branch ? `${subBase}?ref=${encodeURIComponent(branch)}` : subBase;
-      const subResponse = await github.request(subUrl);
-      if (subResponse.status !== 200) {
-        throw new Error(
-          formatGitHubFailure(subResponse, `${owner}/${repo}/${item.path}`, github.hasToken)
-        );
-      }
-      const subContents = parseFetchJson<GitHubContent[]>(subResponse.body);
-      await fs.mkdir(`${destBase}/${item.name}`, { recursive: true });
-      await downloadGitHubDir(
-        subContents,
-        `${destBase}/${item.name}`,
-        owner,
-        repo,
-        branch,
-        fs,
-        github
-      );
-    }
-  }
-}
-
-/**
- * Install a skill from GitHub repository.
- * Tries ZIP-based install first (not rate-limited), falls back to the Contents API.
- */
-async function writeZipFilesToDir(
-  files: Record<string, Uint8Array>,
-  prefix: string,
-  destDir: string,
-  fs: VirtualFS
-): Promise<number> {
-  let fileCount = 0;
-  for (const [path, content] of Object.entries(files)) {
-    if (!path.startsWith(prefix)) continue;
-    const relativePath = path.slice(prefix.length);
-    if (!relativePath || path.endsWith('/')) continue;
-    const filePath = `${destDir}/${relativePath}`;
-    const parentDir = filePath.substring(0, filePath.lastIndexOf('/'));
-    if (parentDir !== destDir) await fs.mkdir(parentDir, { recursive: true });
-    await fs.writeFile(filePath, content);
-    fileCount++;
-  }
-  return fileCount;
-}
-
-async function installFromGitHub(
-  owner: string,
-  repo: string,
-  skillPath: string,
-  skillName: string,
-  fs: VirtualFS,
-  github: GitHubRequestContext,
-  force: boolean = false,
-  fetch?: SecureFetch,
-  branch?: string
-): Promise<{ stdout: string; stderr: string; exitCode: number }> {
-  try {
-    // Check if skill already exists
-    const destDir = `${SKILLS_DIR}/${skillName}`;
-    try {
-      await fs.stat(destDir);
-      if (!force) {
-        return {
-          stdout: '',
-          stderr: `upskill: skill "${skillName}" already exists (use --force to overwrite)\n`,
-          exitCode: 1,
-        };
-      }
-      await fs.rm(destDir, { recursive: true });
-    } catch {
-      // Doesn't exist, continue
-    }
-
-    // Try ZIP-based install first (no rate limit)
-    if (fetch) {
-      const zip = await fetchRepoZip(owner, repo, fetch, branch);
-      if (zip.status === 'ok') {
-        const files = stripZipPrefix(zip.files);
-        const prefix = skillPath.replace(/^\/|\/$/g, '') + '/';
-
-        await fs.mkdir(destDir, { recursive: true });
-        const fileCount = await writeZipFilesToDir(files, prefix, destDir, fs);
-
-        if (fileCount > 0) {
-          await refreshSprinklesAfterInstall();
-          await reloadSkillsAfterInstall();
-          return {
-            stdout: `Installed skill "${skillName}" from ${owner}/${repo}\n`,
-            stderr: '',
-            exitCode: 0,
-          };
-        }
-        // No files found under path — fall through to API
-      }
-    }
-
-    // Fallback: Contents API
-    const base = `https://api.github.com/repos/${owner}/${repo}/contents/${skillPath}`;
-    const url = branch ? `${base}?ref=${encodeURIComponent(branch)}` : base;
-    const response = await github.request(url);
-
-    if (response.status !== 200) {
-      return {
-        stdout: '',
-        stderr: `upskill: ${formatGitHubFailure(response, `${owner}/${repo}/${skillPath}`, github.hasToken)}\n`,
-        exitCode: 1,
-      };
-    }
-
-    const contents = parseFetchJson<GitHubContent[]>(response.body);
-
-    await fs.mkdir(destDir, { recursive: true });
-
-    try {
-      await downloadGitHubDir(contents, destDir, owner, repo, branch, fs, github);
-    } catch (downloadErr) {
-      try {
-        await fs.rm(destDir, { recursive: true });
-      } catch {
-        /* best-effort */
-      }
-      throw downloadErr;
-    }
-
-    await runPostInstallHooks();
-    return {
-      stdout: `Installed skill "${skillName}" from ${owner}/${repo}\n`,
-      stderr: '',
-      exitCode: 0,
-    };
-  } catch (err) {
-    const msg = err instanceof Error ? err.message : String(err);
-    return {
-      stdout: '',
-      stderr: `upskill: failed to install from GitHub: ${msg}\n`,
-      exitCode: 1,
-    };
-  }
-}
-
-/**
- * Parse GitHub repo reference.
- *
- * Accepts either:
- * - bare `owner/repo` or `owner/repo@branch`
- * - full URL `https://github.com/owner/repo[.git][/tree/<branch>[/<subpath>]][/]`
- *
- * For URL form, `/tree/<branch>/<path>` decomposes into `branch` plus an
- * implicit `path`. The caller decides precedence with any explicit `--branch`
- * / `--path` flags.
- *
- * URL form is **https-only and host-anchored** — `http://`, hosts that merely
- * contain `github.com` as a path segment (`evil.com/github.com/...`), and
- * suffix typosquats (`github.com.evil.com`, `github.co`) are rejected.
- */
-export function parseGitHubRef(
-  ref: string
-): { owner: string; repo: string; branch?: string; path?: string } | null {
-  // URL form: https://github.com/owner/repo[.git][/tree/<branch>[/<subpath>]][/]
-  // Anchored: scheme MUST be https; host MUST be exactly `github.com`.
-  const url = ref.match(
-    /^https:\/\/github\.com\/([^/]+)\/([^/]+?)(?:\.git)?(?:\/tree\/([^/]+?)(?:\/(.+?))?)?\/?$/
-  );
-  if (url) {
-    return { owner: url[1], repo: url[2], branch: url[3], path: url[4] };
-  }
-  // Handle owner/repo or owner/repo@branch format
-  const match = ref.match(/^([a-zA-Z0-9_-]+)\/([a-zA-Z0-9_.-]+)(?:@([a-zA-Z0-9_./-]+))?$/);
-  if (match) {
-    return { owner: match[1], repo: match[2], branch: match[3] };
-  }
-  return null;
-}
+// ── Re-exports (preserve the monolith's public surface during the upskill split) ──
+export {
+  _resetBrowseShCatalogCache,
+  _resetGlobalFsCache,
+  fetchBrowseShCatalog,
+  normalizeHostname,
+  parseBrowseShRef,
+  parseGitHubRef,
+  scoreSkills,
+};
 
 type RecommendInstallRecord = { ok: boolean; name: string; error?: string };
 type RepoGroupResult = { errors: string[]; results: RecommendInstallRecord[]; output: string };
@@ -1530,16 +528,6 @@ export async function installRecommendedSkills(
 }
 
 // ── upskill tabs ──
-
-/**
- * Normalize a hostname for catalog matching: lowercases and strips a single
- * leading `www.`. Exported so the browse.sh skill install dispatch can
- * reuse the exact same matching contract this subcommand surfaces to users.
- */
-export function normalizeHostname(host: string): string {
-  const lower = host.toLowerCase();
-  return lower.startsWith('www.') ? lower.slice(4) : lower;
-}
 
 /**
  * Build the install-hint shell line for an origin-advertised upskill rel.
